feat: visualize well intersection in geos-trame (#83)
* feat(wells): add visualization and interaction with well

Improve DeckViewer to be able to show mesh (Region), wells
(InternalWell, Vtkmesh) and perforation (Perforation)

Update also the gui to be able to hide and control radius property of
these datas.

* fix(inspector): fix checking of if an item is renderable or not

* fix(polyline): make sure to generate a polyline based on connectivity

Previously we didn't take in account the connectivty.

The geosDeck used in the test has been updated to check this fix.

* feat(alert) : add alert system

Adding an alert system to notify user about error. It is used in the
viewer.py to notify the user about missing dataset mandaotry to read a
well or a perforation

* fix(IO): support file which doesn't contain an Include tag

* test(well): check well intersection feature for InternalWell andVtkmesh

* test(well) : reuse singlePhaseFlow dataset

To avoid LFS storage limit, test_well_intersection will reuse the mesh
from the other deck xml.